requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
fastapi==0.115.14
google-cloud-firestore==2.21.0
h11==0.16.0
idna==3.10
langchain==0.3.26
langchain-openai==0.3.26
openai==1.92.2
pyjwt==2.10.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
spotipy==2.25.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
